==> 1_preprocessing.ipynb <==
import pandas as pd
import numpy as np

df = pd.read_csv('data/return_predictability_data.csv')

ff_research = pd.read_csv('data/ff_factors.csv')
ff_research.head()

df['Date'] = [''.join(x.split('-')[0:2]) for x in df.DATE]
ff_research['Date'] = ff_research['Date'].apply(str)
df = df.merge(ff_research[['Date', 'macro_mkt-rf', 'macro_hml', 'macro_smb']], on = 'Date', how ='inner' )

df.drop(columns=['name', 'Date'], inplace=True)

df.to_csv('data/features_subset.csv', index=False)
