==> 2_data_cleaning.ipynb <==
import numpy as np
import pandas as pd
import sklearn.preprocessing
import datetime
import pickle

df = pd.read_csv('data/features_subset.csv')

# with open('return_predictability.pkl', 'rb') as f:
#     test = pickle.load(f)

df['DATE'] = pd.to_datetime(df['DATE'])

df = df.loc[df['mvel1']>= 1e+06]

limit = len(df) * .60
df = df.dropna(thresh=limit, axis=1)

pd.set_option('display.max_rows', None)
df.isna().sum().sort_values(ascending=False)

df = df.replace({'sic2': {'missing':'0'}})

df.isna().sum()

date = '1965-01-31'

df['sic2'] = df['sic2'].apply(lambda x: int(x))
df = df.fillna(df.groupby('DATE').transform('median'))
df = df.sort_values(by=['DATE', 'permno'])
# df = df[~(df['DATE'] < '2002-01-31')]
# df = df[~(df['DATE'] < '1990-01-31')]
df = df[~(df['DATE'] < date)]

features = df.columns[~df.columns.isin(['secured',
                                        'aeavol',
                                        'cash',
                                        'chtx',
                                        'cinvest', 'ear',
                                        'nincr',
                                        'roaq', 
                                        'roavol',
                                        'roeq',
                                        'rsup',
                                        'stdacc',
                                        'stdcf',
                                        'ms'
                                        ])].tolist()
df = df[features]

variables = df.columns[~df.columns.isin(['DATE', 'permno'])].tolist()
df[variables] = df[variables].apply(lambda x: x.clip(*x.quantile([0.01, 0.99])))

df.to_csv('data/factors_1965.csv', index=False)
